+++ packages/aws-amplify-react/src/Auth/AuthPiece.jsx
@@ -49,6 +49,16 @@ export default class AuthPiece extends React.Component {
         this.onPhoneNumberChanged = this.onPhoneNumberChanged.bind(this);
     }
 
+    componentDidMount() {
+        if (window && window.location && window.location.search) {
+            if (!this.props.authData || !this.props.authData.username) {
+                const searchParams = new URLSearchParams(window.location.search);
+                const username = searchParams ? searchParams.get('username') : undefined;
+                this.setState({ username });
+            }
+        }
+    }
+
     getUsernameFromInput() {
         const { usernameAttributes = 'username' } = this.props;
         switch(usernameAttributes) {
@@ -57,7 +67,7 @@ export default class AuthPiece extends React.Component {
             case UsernameAttributes.PHONE_NUMBER:
                 return this.phone_number;
             default:
-                return this.inputs.username;
+                return this.inputs.username || this.state.username;
         }
     }
 
@@ -91,6 +101,7 @@ export default class AuthPiece extends React.Component {
                 <FormField theme={theme}>           
                     <InputLabel theme={theme}>{I18n.get(this.getUsernameLabel())} *</InputLabel>
                     <Input
+                        defaultValue={this.state.username}
                         autoFocus
                         placeholder={I18n.get('Enter your username')}
                         theme={theme}