+++ packages/core/src/Amplify.ts
@@ -19,6 +19,7 @@ export default class Amplify {
     static Pushnotification = null;
     static UI = null;
     static XR = null;
+    static Predictions = null;
 
     static Logger = LoggerClass;
     static ServiceWorker = null;