+++ packages/aws-amplify-react-native/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.16](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.1.15...aws-amplify-react-native@2.1.16) (2019-07-30)
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+
+
+
+
+## [2.1.15](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.1.14...aws-amplify-react-native@2.1.15) (2019-07-18)
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+
+
+
+
+<a name="2.1.14"></a>
+## [2.1.14](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.1.14-unstable.0...aws-amplify-react-native@2.1.14) (2019-07-09)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
 <a name="2.1.14-unstable.0"></a>
 ## [2.1.14-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.1.13...aws-amplify-react-native@2.1.14-unstable.0) (2019-06-27)